requirements.txt
# Core ML
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
timm>=0.9.0
opencv-python>=4.8.0
joblib>=1.3.0

# Web (Flask stack)
Flask>=3.0.0
Werkzeug>=3.0.0
Jinja2>=3.1.2
python-dotenv>=1.0.1

# Security helpers
Flask-Login>=0.6.3
Flask-Limiter>=3.5.0
Flask-Talisman>=1.1.0
Flask-WTF>=1.2.1
Flask-Bcrypt>=1.0.1
cryptography>=42.0.0
firebase-admin>=6.3.0

# Utilities
numpy>=1.26.0
matplotlib>=3.9.0
transformers>=4.38.0